+++ run_ex.py
@@ -3,59 +3,48 @@ from concurrent.futures import ProcessPoolExecutor
 from multiprocessing import Manager
 import concurrent.futures
 from tqdm import tqdm
-import gc
-import random
-import time
 from src.dataset import Dataset
 from src.model import Model
 
 
 def worker(config, gpu_queue):
-    gc.collect()
-
     gpu_id = gpu_queue.get()
 
     try:
-        while True:
-            ds = Dataset(data_type="experiment", n_train_samples=config['train_size'], test_samples=["4", "22"],
-                         seed=config['run'])
-            x_train, y_train = ds.prepare_data(ds.train_indices)
-            x_val, y_val = ds.prepare_data(ds.test_indices)
-            input_shape = (x_train.shape[1], x_train.shape[2])
+        ds = Dataset(data_type="experiment", n_train_samples=config['train_size'], seed=config['run'])
+        x_train, y_train = ds.prepare_data(ds.train_indices)
+        x_val, y_val = ds.prepare_data(ds.test_indices)
+        input_shape = (x_train.shape[1], x_train.shape[2])
 
-            if 'theta' not in config:
-                config['theta'] = 0.04 # 0.03866414715431152
+        if 'theta' not in config:
+            config['theta'] = 1000
 
-            if 'dropout_rate' not in config:
-                config['dropout_rate'] = 0.25
+        if 'tgds_ratio' not in config:
+            config['tgds_ratio'] = 0.6
 
-            if 'tgds_ratio' not in config:
-                config['tgds_ratio'] = 0.5
+        if 'loss_tolerance' not in config:
+            config['loss_tolerance'] = 0
 
-            model = Model(
-                data_type=ds.data_type,
-                loss_function=config['loss'],
-                scaler=ds.scaler,
-                steepness=ds.steepness,
-                input_shape=input_shape,
-                neurons=32,
-                learning_rate=0.001,
-                theta=config['theta'],
-                tgds_ratio=config['tgds_ratio'],
-                batch_size=128,
-                verbose=False,
-                dropout_rate=config['dropout_rate'],
-                n_lstm_layers=3,
-                seed=config['run'],
-                gpu=int(gpu_id),
-                loss_normalized=True,
-                ex_name="dropout_check"
-            )
-            model.train(x_train, y_train, x_val=x_val, y_val=y_val, epochs=200, check=False)
-            model.predict(dataset=ds, show_plot=False)
-            is_deleted = model.get_result(check=False)
-            if not is_deleted:
-                break
+        if 'steepness' not in config:
+            config['steepness'] = ds.steepness
+
+        model = Model(
+            data_type=ds.data_type,
+            n_train_samples=config['train_size'],
+            loss_function=config['loss'],
+            scaler=ds.scaler,
+            steepness=config['steepness'],
+            input_shape=input_shape,
+            theta=config['theta'],
+            tgds_ratio=config['tgds_ratio'],
+            verbose=False,
+            seed=config['run'],
+            gpu=int(gpu_id),
+            loss_tolerance=config['loss_tolerance'],
+            ex_name=config['ex_name'],
+        )
+        model.train(x_train, y_train, x_val=x_val, y_val=y_val, epochs=300)
+        model.predict(dataset=ds, show_plot=False)
     except Exception as e:
         print(str(e))
         traceback.print_exc()
@@ -63,18 +52,68 @@ def worker(config, gpu_queue):
     gpu_queue.put(gpu_id)
 
 
+thetas = [0.001, 0.01, 0.1, 1, 10, 100, 1000, 10000]
+tgds_ratios = [0.001, 0.01, 0.1, 1, 10, 100, 1000, 10000]
+train_sizes = [1, 2, 3, 4, 5, 6]
+losses = ['mean_squared_error', 'soft_weighted_two_state']
+loss_tolerances = [0, 0.01, 0.1]
+incorrect_steepnesses = [6.388 / 100, 6.388 / 2, 6.388, 6.388 * 2, 6.388 * 100]
+
 configs = []
 for run in range(5):
-    for loss in ['two_state', 'mean_squared_error']:  # ['two_state', 'mean_squared_error']:
-        for train_size in [2, 6]:
-            for dropout_rate in [0, 0.25]:
-                configs.append({'run': run, 'train_size': train_size, 'loss': loss, 'dropout_rate': dropout_rate})
-random.shuffle(configs)
-configs = sorted(configs, key=lambda x: x['run'])
+
+    # Incorrect prior knowledge
+    for st in incorrect_steepnesses:
+        configs.append({
+            'run': run,
+            'train_size': 4,
+            'loss': 'soft_weighted_two_state',
+            'steepness': st,
+            'ex_name': 'incorrect_prior_knowledge'
+        })
+
+    # Different train sizes with different loss tolerances
+    for train_size in train_sizes:
+        for tolerance in loss_tolerances:
+            configs.append({
+                'run': run,
+                'train_size': train_size,
+                'loss': 'soft_weighted_two_state',
+                'loss_tolerance': tolerance,
+                'ex_name': 'diff_sizes'
+            })
+        configs.append({
+            'run': run,
+            'train_size': train_size,
+            'loss': 'mean_squared_error',
+            'ex_name': 'diff_sizes'
+        })
+
+    # Sensitivity analysis for lambda
+    for theta in thetas:
+        configs.append({
+            'run': run,
+            'train_size': 4,
+            'loss': 'soft_weighted_two_state',
+            'theta': theta,
+            'tgds_ratio': 1,
+            'ex_name': 'sensitivity_analysis_lambda'
+        })
+
+    # Sensitivity analysis for tgds_ratio
+    for tgds_ratio in tgds_ratios:
+        configs.append({
+            'run': run,
+            'train_size': 4,
+            'loss': 'soft_weighted_two_state',
+            'theta': 1000,
+            'tgds_ratio': tgds_ratio,
+            'ex_name': 'sensitivity_analysis_tgds_ratio'
+        })
 
 
 def main():
-    gpus = ["0", "1", "0", "1", "0", "1"]
+    gpus = ["0", "1", "2", "3"]
 
     with Manager() as manager:
         gpu_queue = manager.Queue()
